+++ client/src/main.tsx
@@ -4,7 +4,7 @@ import App from "./App.tsx";
 import "./index.css";
 import { BrowserRouter } from "react-router-dom";
 import Header from "./components/Header.tsx";
-import { Footer } from "./components/Home/Footer.tsx";
+import { Footer } from "./components/Footer/Footer.tsx";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
