--- client/src/pages/Departments.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Departments = () => {
-  return (
-    <div>Departments</div>
-  )
-}
-
-export default Departments
